Add aux method for editing
- Modify modal to edit the values in the project or to-do
