+++ src/assets/stylesheets/style.scss
@@ -98,6 +98,10 @@ nav {
 
       .title-btn {
         @include flex(row, space-between, center);
+
+        &:hover {
+          cursor: pointer;
+        }
       }
     }
   }
